# server/openapi.yaml
openapi: 3.1.0
info:
  title: Sahayak API
  version: 1.0.0
  description: PIN-authenticated API for hirers and labourers. Phone ownership is not verified.
servers:
  - url: http://localhost:8080
paths:
  /health:
    get:
      operationId: health
      responses:
        "200": {description: Healthy, content: {application/json: {schema: {$ref: "#/components/schemas/Health"}}}}
  /v1/auth/register:
    post:
      operationId: register
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/RegisterRequest"}}}
      responses:
        "201": {$ref: "#/components/responses/AuthSuccess"}
        "400": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
  /v1/auth/login:
    post:
      operationId: login
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/LoginRequest"}}}
      responses:
        "200": {$ref: "#/components/responses/AuthSuccess"}
        "401": {$ref: "#/components/responses/Error"}
        "429": {$ref: "#/components/responses/Error"}
  /v1/auth/refresh:
    post:
      operationId: refreshSession
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/RefreshRequest"}}}
      responses:
        "200": {$ref: "#/components/responses/AuthSuccess"}
        "401": {$ref: "#/components/responses/Error"}
  /v1/auth/logout:
    post:
      operationId: logout
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/RefreshRequest"}}}
      responses:
        "204": {description: Session revoked}
  /v1/me:
    get:
      operationId: getMe
      security: [{accessBearer: []}]
      responses:
        "200": {description: Current user, content: {application/json: {schema: {$ref: "#/components/schemas/User"}}}}
        "401": {$ref: "#/components/responses/Error"}
  /v1/jobs/{jobId}/payment:
    parameters: [{$ref: "#/components/parameters/JobID"}]
    get:
      operationId: getJobPayment
      security: [{accessBearer: []}]
      responses:
        "200": {description: Payment, content: {application/json: {schema: {$ref: "#/components/schemas/Payment"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
  /v1/jobs/{jobId}/payment/mark-paid:
    parameters: [{$ref: "#/components/parameters/JobID"}]
    post:
      operationId: markJobPaymentPaid
      security: [{accessBearer: []}]
      responses:
        "200": {description: Updated payment, content: {application/json: {schema: {$ref: "#/components/schemas/Payment"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
components:
  securitySchemes:
    accessBearer: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    JobID: {name: jobId, in: path, required: true, schema: {type: string, format: uuid}}
  responses:
    AuthSuccess:
      description: Authenticated session
      content: {application/json: {schema: {$ref: "#/components/schemas/AuthResponse"}}}
    Error:
      description: Error response
      content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}
  schemas:
    Health:
      type: object
      required: [status]
      properties: {status: {type: string, const: ok}}
    Role: {type: string, enum: [HIRER, LABOURER]}
    Phone:
      type: string
      pattern: '^\+91[6-9][0-9]{9}$'
    Pin:
      type: string
      pattern: '^[0-9]{4}$'
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [phoneNumber, pin, role]
      properties:
        phoneNumber: {type: string}
        pin: {$ref: "#/components/schemas/Pin"}
        role: {$ref: "#/components/schemas/Role"}
    LoginRequest:
      type: object
      additionalProperties: false
      required: [phoneNumber, pin]
      properties:
        phoneNumber: {type: string}
        pin: {$ref: "#/components/schemas/Pin"}
    RefreshRequest:
      type: object
      additionalProperties: false
      required: [refreshToken]
      properties: {refreshToken: {type: string, minLength: 1}}
    User:
      type: object
      required: [id, phoneNumber, role, createdAt, updatedAt]
      properties:
        id: {type: string, format: uuid}
        phoneNumber: {$ref: "#/components/schemas/Phone"}
        role: {$ref: "#/components/schemas/Role"}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    AuthResponse:
      type: object
      required: [user, accessToken, refreshToken, expiresIn]
      properties:
        user: {$ref: "#/components/schemas/User"}
        accessToken: {type: string}
        refreshToken: {type: string}
        expiresIn: {type: integer, const: 900}
    Payment:
      type: object
      required: [id, jobId, amountPaise, currency, status, createdAt, updatedAt]
      properties:
        id: {type: string, format: uuid}
        jobId: {type: string, format: uuid}
        amountPaise: {type: integer, format: int64, minimum: 1}
        currency: {type: string, const: INR}
        status: {type: string, enum: [PENDING, PAID_OFFLINE]}
        paidAt: {type: [string, "null"], format: date-time}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
            retryAt: {type: string, format: date-time}